requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-filter==23.3
django-cors-headers==4.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
celery==5.3.1
redis==5.0.1
drf-yasg==1.21.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1
openpyxl==3.1.2
reportlab==4.0.8
xlwt==1.3.0